requirements.txt
Babel==2.7.0
CacheControl==0.11.7
cachetools==3.1.1
cbor==1.0.0
Cerberus==1.3.1
certifi==2019.3.9
chardet==3.0.4
Click==7.0
dominate==2.3.5
Flask==1.0.3
Flask-Cors==3.0.7
Flask-SQLAlchemy==2.4.0
Flask-SQLAlchemy-Session==1.1
functoolz==0.4
GeoAlchemy2==0.6.2
geojson==2.4.1
httplib2==0.12.3
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
Khayyam==3.0.17
MarkupSafe==1.1.1
ntplib==0.3.3
oauth2==1.9.0.post1
oauthlib==3.0.1
psycopg2-binary==2.8.2
pyparsing==2.4.0
python-dotenv==0.10.3
python-rapidjson==0.7.1
pytz==2019.1
PyYAML==5.1
redis==3.2.1
requests==2.22.0
Shapely==1.6.4.post2
six==1.12.0
SQLAlchemy==1.3.3
SQLAlchemy-Utils==0.33.11
suds-jurko==0.6
swaggertools==0.0.4
toolz==0.9.0
typing==3.6.6
u-msgpack-python==2.5.1
urllib3==1.25.3
Werkzeug==0.15.4
